Extract CSS compatibility processor
